+++ test/axios2.test.js
@@ -35,7 +35,8 @@ let {
     BASE_URL_FOR_ST,
     getNumberOfTimesRefreshAttempted,
     coreTagEqualToOrAfter,
-    checkIfJWTIsEnabled
+    checkIfJWTIsEnabled,
+    resetAuthHttpRequestFetch
 } = require("./utils");
 const { spawn } = require("child_process");
 let { ProcessState, PROCESS_STATE } = require("../lib/build/processState");
@@ -64,7 +65,7 @@ describe("Axios AuthHttpRequest class tests", function() {
     });
 
     beforeEach(async function() {
-        AuthHttpRequestFetch.initCalled = false;
+        resetAuthHttpRequestFetch();
         global.document = {};
         ProcessState.getInstance().reset();
         let instance = axios.create();