+++ test/cookie.handler.test.js
@@ -15,7 +15,7 @@
 let jsdom = require("mocha-jsdom");
 let AuthHttpRequest = require("../index.js").default;
 let { default: AuthHttpRequestFetch } = require("../lib/build/fetch");
-let { BASE_URL, BASE_URL_FOR_ST, startST } = require("./utils");
+let { BASE_URL, BASE_URL_FOR_ST, startST, resetAuthHttpRequestFetch } = require("./utils");
 let { default: CookieHandlerReference } = require("../lib/build/utils/cookieHandler/index");
 const { spawn } = require("child_process");
 let axios = require("axios");
@@ -49,7 +49,7 @@ describe("Cookie Handler Tests", function() {
     beforeEach(async function() {
         consoleLogs = [];
         CookieHandlerReference.instance = undefined;
-        AuthHttpRequestFetch.initCalled = false;
+        resetAuthHttpRequestFetch();
         global.document = {};
         ProcessState.getInstance().reset();
         let instance = axios.create();