+++ test/fetch.test.js
@@ -32,7 +32,8 @@ let {
     BASE_URL_FOR_ST,
     coreTagEqualToOrAfter,
     checkIfJWTIsEnabled,
-    addBrowserConsole
+    addBrowserConsole,
+    resetAuthHttpRequestFetch
 } = require("./utils");
 const { spawn } = require("child_process");
 let { ProcessState, PROCESS_STATE } = require("../lib/build/processState");
@@ -74,7 +75,7 @@ describe("Fetch AuthHttpRequest class tests", function() {
     });
 
     beforeEach(async function() {
-        AuthHttpRequestFetch.initCalled = false;
+        resetAuthHttpRequestFetch();
         global.document = {};
         ProcessState.getInstance().reset();
         let instance = axios.create();