+++ test/window.handler.test.js
@@ -14,8 +14,7 @@
  */
 let jsdom = require("mocha-jsdom");
 let AuthHttpRequest = require("../index.js").default;
-let { default: AuthHttpRequestFetch } = require("../lib/build/fetch");
-let { BASE_URL, BASE_URL_FOR_ST, startST } = require("./utils");
+let { BASE_URL, BASE_URL_FOR_ST, startST, resetAuthHttpRequestFetch } = require("./utils");
 let { default: WindowHandlerReference } = require("../lib/build/utils/windowHandler/index");
 const { spawn } = require("child_process");
 let axios = require("axios");
@@ -49,7 +48,7 @@ describe("Window handler tests", function() {
     beforeEach(async function() {
         consoleLogs = [];
         WindowHandlerReference.instance = undefined;
-        AuthHttpRequestFetch.initCalled = false;
+        resetAuthHttpRequestFetch();
         global.document = {};
         ProcessState.getInstance().reset();
         let instance = axios.create();